feat: Mute word audio on matching until matched (configurable)
Because providing audio on words in matching exercises
makes it too tempting to just rely on the audio of a word
instead of spending the necessary energy to recall meaning,
we need to mute audio from words in matching exercises
until a word and a picture/symbol (with audio) has been matched.

After a match, the audio should reinforce memorization with context
and possibly also feel like a reward.

If the audio on the word is available before a match has been achieved,
it may bee too tempting to just listen to the audio of the word
and not spend effort recalling the meaning of the word from memory,
and therefore gain much less memory retention training
from the matching exercise.

We make this new feature configurable by content developers,
so that this feature can be turned on and off on a unit by unit basis.

This commit will:
- mute audio on words in matching exercises
  if the word has not yet been matched to a picture/symbol
  and also the unit has NOT been configured by a content developer
  to override muting (i.e. the default will be to mute)

**Certification**
- [X] I certify that <!-- Check the box to certify: [X] -->
- I have read the [contributing guidelines](
  https://github.com/nodepa/seedlingo/blob/main/.github/CONTRIBUTING.md)
- I license these contributions to the public under Seedlingo's
  [LICENSE](https://github.com/nodepa/seedlingo/blob/main/LICENSE.md)
  and have the rights to do so.

Signed-off-by: toshify <4579559+toshify@users.noreply.github.com>