test: Update test depending on translated content
Because content has been translated into English,
but an end-to-end test of the review system
evaluates the presence of the first two words of the first unit,
failing the test because those words are now in English.

This commit will update the relevant test
to check for the English translation instead.

**Certification**
- [X] I certify that <!-- Check the box to certify: [X] -->
- I have read the [contributing guidelines](
  https://github.com/nodepa/seedlingo/blob/main/.github/CONTRIBUTING.md)
- I license these contributions to the public under Seedlingo's
  [LICENSE](https://github.com/nodepa/seedlingo/blob/main/LICENSE.md)
  and have the rights to do so.

Signed-off-by: toshify <4579559+toshify@users.noreply.github.com>