fix: Amend resize of answer option text
When creating content in a language that uses a writing system
with an alphabet script rather than a logography,
answer options tend to consist of more symbols.
This surfaces a need to handle scaling of both alpabetic scripts
(which tend to form words that are significantly wider than tall,
in particular in languages that use inflection or compounding)
and logogrphic scrips (which can form words that are taller than wide,
i.e. one character words).
Since CSS does not allow us to simply define that a paragraph should
fit the container it is placed in,
i.e. scale up to fill the container if the text is one word
or scale down to fit within the container if the text is very long,
we need sensible parameters for font size to:
- display short text filling at least 50% of its container
- display long text large enough to be legible
  while not overflowing its container unless it becomes illegible

This commit will:
- amend font-size styles on answer buttons

**Certification**
- [X] I certify that <!-- Check the box to certify: [X] -->
- I have read the [contributing guidelines](
  https://github.com/nodepa/seedlingo/blob/main/.github/CONTRIBUTING.md)
- I license these contributions to the public under Seedlingo's
  [LICENSE](https://github.com/nodepa/seedlingo/blob/main/LICENSE.md)
  and have the rights to do so.

Signed-off-by: toshify <4579559+toshify@users.noreply.github.com>